requirements.txt
# Основные зависимости из image-dedup
numpy<2.0.0
mpmath>=1.3.0
open-clip-torch>=2.23.0
SQLAlchemy>=2.0.0
psycopg[binary,pool]>=3.1.0
pgvector>=0.2.0
typer[all]>=0.12.3
PyQt6>=6.6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
faiss-cpu>=1.7.4
networkx>=3.2.1
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.66.0

# Зависимости для Torch с указанием источника (для установки вручную)
# pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu121
# torch==2.1.0
# torchvision==0.16.0

# Зависимости из exifreader
ExifRead==3.5.1

# Зависимости для тестов находятся в requirements-dev.txt